Fix Query() returning repeater's address
fixes an issue that happens when a mDNS query
is broadcast and a repeater answers instead or
before the real host. In this situation,
the original code wrongly returns the address
of the repeater.

The new code actually parses the DNS
response UDP packet to find the answer.
This works with and without repeaters.