Set zone for IPv6 ULA addresses
`addrWithOptionalZone` only adds the name of the zone when the IPv6
address is link-local unicast or broadcast. This causes tests to fail
when an IPv6 Unique Local Address (ULA) address (`fc00::/7`) is chosen.

ULA addresses are the moral equivalent of IPv4 private RFC1918
addresses, so it makes sense to have mDNS zones set for these IPs.

There appears to be a mismatch in the expectations of the test and the
implementation: The implementation chooses the first real IPv6 address
it receives from the IPs assigned to the interface while the test code
expects a Zone to be set, which only happens for link-local (and now
ULA) addresses. This will be addressed in a follow-up bug/PR.

For now, this PR enables tests to run in environments where the selected
address is a ULA address (such as my personal computer and the docker
containers used for testing).